Features/form enfants

+++ src/Controller/ProfilController.php
@@ -14,17 +14,14 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 class ProfilController extends AbstractController
 {
-    /**
-     * @Route("parents/profil", name="profil")
+     * @Route("/parents/profil", name="profil")
      */
     public function index(AuthorizationCheckerInterface $authChecker)
     {
-//        return $this->render('403/403.html.twig',[
-//                'erreur' => 'ACCES INTERDIT'
-//            ]);
         if (true === $authChecker->isGranted('ROLE_PARENT'))
         {
             $parent = $this->getUser();
@@ -66,39 +63,67 @@ class ProfilController extends AbstractController
     }
 
     /**
-     * @Route("/profil/edit/{id}", name="parents_edit", methods={"GET","POST"})
+     * @Route("/parents/profil/edit", name="parents_edit", methods={"GET","POST"})
      */
-    public function editParents(Request $request, Parents $parent): Response
+    public function editParent(Request $request, AuthorizationCheckerInterface $authChecker): Response
     {
-        $form = $this->createForm(ParentsType::class, $parent);
-        $form->handleRequest($request);
+        if (true === $authChecker->isGranted('ROLE_PARENT')) {
+            $user = $this->getUser();
+            $form = $this->createForm(ParentsType::class, $user);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $user->setParentsUpdatedAt(new \DateTime('now'));
+                $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('profil', [
-                'id' => $parent->getId(),
+                return $this->redirectToRoute('profil');
+            }
+
+            return $this->render('profilParents/edit.html.twig', [
+                'parent' => $user,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            return $this->render('403/403.html.twig', [
+                'erreur' => 'ACCES FORBIDEN'
             ]);
         }
-
-        return $this->render('profilParents/edit.html.twig', [
-            'parent' => $parent,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
-     * @Route("/profil/delete/{id}", name="parents_delete", methods={"DELETE"})
+     * @Route("/parents/profil/delete/{id}", name="parents_delete", methods={"DELETE"})
      */
-    public function delete(Request $request, Parents $parent): Response
+    public function deleteParent(Request $request, Parents $parent, TokenStorageInterface $tokenStorage, AuthorizationCheckerInterface $authChecker): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $parent->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($parent);
-            $entityManager->flush();
-        }
-        SecurityController::logout();
+        if (true === $authChecker->isGranted('ROLE_PARENT')) {
+                if ($this->isCsrfTokenValid('delete' . $parent->getId(), $request->request->get('_token'))) {
+                    $entityManager = $this->getDoctrine()->getManager();
 
-        return $this->redirectToRoute('accueil');
+                    $enfants = $this->getUser()->getChildren();
+                    $i = 0;
+
+                    while ($enfants[$i] != Null) {
+                        $o = 0;
+                        $maladie = $enfants[$i]->getDiseases();
+
+                        while ($maladie[$o] != Null) {
+                            $entityManager->remove($maladie[$o]);
+                            $o++;
+                        }
+                        $entityManager->remove($enfants[$i]);
+                        $i++;
+                    }
+
+                    $this->get('security.token_storage')->setToken(null);
+                    $entityManager->remove($parent);
+                    $entityManager->flush();
+
+                    return $this->redirectToRoute('accueil');
+                } else {
+                    return $this->render('403/403.html.twig', [
+                        'erreur' => 'ACCES FORBIDEN'
+                    ]);
+                }
+        }
     }
 }
